// packages/trading50/src/screens/types.ts
import { TextStyle, ViewStyle } from "react-native";
import { TObject } from "./model";
export type TVoidCalBack = () => void;
export type TStringCallback = (str: string) => void;
export type TErrorCallback = (str: string) => void;
export type TSuccessCallback = (obj: Object) => void;
export type TAnyCallback = (obj: any) => void;
export type TKeyText = {
  key: string;
  text: string;
};
export type TProps = {
  navigation?: any;
  route?: any;
  children?: any;
  userToken?: string | null;
  overrideStyle?: ViewStyle;
  onPress?: any;
  style?: ViewStyle | TextStyle;
  pstyle?: ViewStyle | TextStyle; // mainly used for top level elemnet
  theme?: string;
  primary?: any;
  secondary?: any;
  center?: any;
  item?: any;
  text?: string;
  icon?: string;
  loading?: boolean;
  size?: number;
  color?: string;
  items?: Array<TObject | TKeyText>;
  objArray?: Array<TObject>;
  object?: TObject;
  numberOfLines?: number;
  visible?: boolean;
  onSearch?: TStringCallback;
  onChangeText?: TStringCallback;
  title?: string;
  subtitle?: string;
  placeholder?: string;
  onPressRightIcon?: TVoidCalBack;
  onPressLeftIcon?: TVoidCalBack;
  value?: any;
  onOk?: TVoidCalBack;
  onCancel?: TVoidCalBack;
  onChange?: TAnyCallback;
  onSignOut?: TVoidCalBack;
};